Add files via upload
This is my very first and very own Python program :) I had fun creating it, but the input verification was more challenging than I thought.
I finished day 7 of 100 in the '100 Days of Code Python Bootcamp', so I'm pretty happy with the results. However, I want to polish the code and add some graphics in the future. And  I need separate windows for the available networks list. Currently, I'm calculating up to 10 subnets in class C only because of the limited space on the screen.